Move test_shel_jobs to res/fm/test_shell
