+++ tests/libres_tests/res/fm/test_shell.py
@@ -5,6 +5,7 @@ import subprocess
 
 import pytest
 
+from ert._c_wrappers.enkf import ResConfig
 from ert._c_wrappers.fm.shell import (
     careful_copy_file,
     copy_directory,
@@ -353,3 +354,57 @@ FORWARD_MODEL DELETE_DIRECTORY(<DIRECTORY>=mydir)
         assert os.path.exists("realization-0/iter-0/copied3.txt")
         assert not os.path.exists("realization-0/iter-0/mydir")
         assert os.path.exists("realization-0/iter-0/mydir2")
+
+
+@pytest.fixture
+def minimal_case(tmpdir):
+    with tmpdir.as_cwd():
+        with open("config.ert", "w") as fout:
+            fout.write("NUM_REALIZATIONS 1")
+        yield
+
+
+def test_shell_script_jobs_availability(minimal_case):
+    res_config = ResConfig("config.ert")
+    site_config = res_config.site_config
+
+    installed_jobs = site_config.get_installed_jobs()
+    fm_shell_jobs = {}
+    for job in installed_jobs:
+        exe = job.get_executable()
+        if "shell_scripts" in exe:
+            fm_shell_jobs[job.name().upper()] = exe
+
+    list_from_content = res_config.ert_workflow_list
+    wf_shell_jobs = {}
+    for wf_name in list_from_content.getJobNames():
+        exe = list_from_content.getJob(wf_name).executable()
+        if exe and "shell_scripts" in exe:
+            wf_shell_jobs[wf_name] = exe
+
+    assert fm_shell_jobs == wf_shell_jobs
+
+
+def test_shell_script_jobs_names(minimal_case):
+    shell_job_names = [
+        "DELETE_FILE",
+        "DELETE_DIRECTORY",
+        "COPY_DIRECTORY",
+        "MAKE_SYMLINK",
+        "MOVE_FILE",
+        "MAKE_DIRECTORY",
+        "CAREFUL_COPY_FILE",
+        "SYMLINK",
+        "COPY_FILE",
+    ]
+
+    res_config = ResConfig("config.ert")
+    found_jobs = set()
+    list_from_content = res_config.ert_workflow_list
+    for wf_name in list_from_content.getJobNames():
+        exe = list_from_content.getJob(wf_name).executable()
+        if exe and "shell_scripts" in exe:
+            assert wf_name in shell_job_names
+            found_jobs.add(wf_name)
+
+    assert len(shell_job_names) == len(found_jobs)
